Fix OSS build
