+++ dgraph/cmd/live/run.go
@@ -697,8 +697,11 @@ func run() error {
 	}
 
 	creds := z.NewSuperFlag(Live.Conf.GetString("creds")).MergeAndCheckDefault(x.DefaultCreds)
+	keys, err := ee.GetKeys(Live.Conf)
+	if err != nil {
+		return err
+	}
 
-	var err error
 	x.PrintVersion()
 	opt = options{
 		dataFiles:       Live.Conf.GetString("files"),
@@ -717,6 +720,7 @@ func run() error {
 		ludicrousMode:   Live.Conf.GetBool("ludicrous"),
 		upsertPredicate: Live.Conf.GetString("upsertPredicate"),
 		tmpDir:          Live.Conf.GetString("tmp"),
+		key:             keys.EncKey,
 	}
 
 	forceNs := Live.Conf.GetInt64("force-namespace")
@@ -737,12 +741,6 @@ func run() error {
 	}
 
 	z.SetTmpDir(opt.tmpDir)
-
-	keys, err := ee.GetKeys(Live.Conf)
-	if err != nil {
-		return err
-	}
-	opt.key = keys.EncKey
 
 	go func() {
 		if err := http.ListenAndServe(opt.httpAddr, nil); err != nil {
